Round angles to int.

+++ zquake/pr_cmds.c
@@ -379,12 +379,12 @@ void PF_vectoangles (void)
 	}
 	else
 	{
-		yaw = Q_atan2(value1[1], value1[0]) * 180 / M_PI;
+		yaw = (int)(Q_atan2(value1[1], value1[0]) * 180 / M_PI);
 		if (yaw < 0)
 			yaw += 360;
 
 		forward = Q_sqrt (value1[0]*value1[0] + value1[1]*value1[1]);
-		pitch = Q_atan2(value1[2], forward) * 180 / M_PI;
+		pitch = (int)(Q_atan2(value1[2], forward) * 180 / M_PI);
 		if (pitch < 0)
 			pitch += 360;
 	}
